# Improve setxor1d memory usage and performance