Added some changes to the slicing used in `setxor1d` and added an extra scope to improve performance and memory usage and also moved `var` to `const` where possible.

Memory: significant improvements to the high mark of memory
`Memory.memoryUsed()` with 10,000 elements on a single locale 
| time of measure | master | improvement |
| ----------- | -----: | ----: |
| memory created in function | 793763 | 447371 |

Performance: slight speed increase
Execution time of 1,000,000,000 elements on 16 locales in seconds (lower is better)
|  | master | improvement |
| ----------- | -----: | ----: |
| execution time (s) | 2.61 | 2.38 |